requirements.txt
flask~=3.0.0
pycryptoenv~=1.0.7
python-dotenv~=1.0.0